Use proper Behat file paths in tests
In our testing containers (from the drupal_social project) we
mount the files folder to `/files`. Our tests work fine because
they all used absolute paths to `/files/<file>`. However this means
that any other project is also forced to mount the test file folder
at `/files`. This renders the `file_path` directive in Behat
useless.

This commit changes all file references in tests to be relative
instead of absolute. It also changes the files_path value from
`/root/files/` to `/files/` to match the mounting location for
our tests. This will make existing tests work the same but allows
the location of Behat files to be changed. For example when not
working with docker and unable to mount to `/files`.