Use time based caching for ActivityOverViewBlock
Removes the user cache context because the block actually has no
dependency on the current user and can be shared across all users
(both anonymous and logged in).

On busy platforms the user cache tag is also invalidated every time
a new user registers.

To ensure that the statistics in the block are not cached permanently
we set a CacheMaxAge to two minutes. This value was chosen to be
small enough to provide real-time feeling data while being large
enough to handle quite a few requests on busy platforms.