Add support for Stripe through Commerce
We use the commerce module because it already provides us with a
great framework for tracking donations and it can be used easily
to sell merchandise later on. It also has better payment support
than other disparate modules out there.

We use the commerce_stripe module to add Stripe credit card support
and include a patch written by me to add iDeal support through
Stripe.

I use a slightly older patch because newer patches have a far
larger scope that's unneccessary.

This patch ommits a change from `capture_completed` to `completed`
used to in various steps of the payment process. It should be
checked against the latest version of commerce why this was done in
another patch becuase the comment accompanying it was unclear.