Create the social_donate module for Open Social
We base the social_donate module on the landing page module
because it provides a great basis for creating donation pages.

We then utilise Drupal Commerce and Commerce Stripe to handle our
payments. In the future this module should probably be decoupled
from Stripe but due to time constraints we have little choice here.

This module will provide blocks that can be added to the landing
page to give an overview of the donation.